refactor(js): split code into page specific modules
